examples/google-adk-agents/requirements.txt
google-adk>=1.5.0
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv 